feat: add probes for the agent and deployment
# Why

The platform & agents currently do not register any health
endpoints with Kubernetes causing broken systems to exist
undetected.

# What

Register (optional) startup and liveness probes. For the
agent we do not offer startup probes out of the box, because
these are only supported in later Kubernetes versions and
we want to reduce the likelyhood of upgrade problems.