fix: outpost - inline actions for old helm builds and add podLabels
Older versions of Go [do not support multiline action invocations in templates](https://github.com/golang/go/issues/29770). It looks like the version of helm that is packaged with our spinnaker instance predates that fix. This change only refactors template actions in the outpost chart, so there may be other multiline actions elsewhere.

This change also adds support for a podLabels value (analogous to podAnnotations) so that we can set some labels only on the statefulset pods.